Fix,Rename: 데이터 로직 변경 및 Json 함수 폴더 변경
