+++ src/functions/post.js
@@ -1,7 +1,11 @@
 // * 포스트 상세보기 가져오는 함수
 function selectPosts(data, parsedUrl) {
-  let inputId = parsedUrl.charAt(parsedUrl.length - 1);
-  let selectedPost = data[inputId - 1];
+  let inputId = parsedUrl.charAt(parsedUrl.length - 1); // id값 뽑아내기
+  /**
+   * TODO. 2025-02-25 <데이터 로직 변경>
+   * TODO. 삭제 기능 추가로 인해 data[inputId - 1] 가 아닌, find()를 이용해 선택한 포스트의 아이디값과 일치하는 데이터 가져오기
+   */
+  let selectedPost = data.find(obj => obj.id == inputId);
   let htmlPage = `
   <!DOCTYPE html>
 <html lang="en">
